+++ lua/haskell-tools/hoogle/util.lua
@@ -93,4 +93,12 @@ function M.mk_hoogle_entry(data)
   }
 end
 
+function M.merge_telescope_opts(opts)
+  local default_layout = {
+    layout_strategy = 'horizontal',
+    layout_config = { preview_width = 80 },
+  }
+  return vim.tbl_extend('force', default_layout, opts or {})
+end
+
 return M