+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Pass the custom options to hoogle telescope, so that users' custom telescope
+  themes, etc. can be supported.
 
 ## [1.4.3] - 2022-12-06
 ### Fixed